refactor(service, PairMatcher): service 클래스 도입, 기존 컨트롤러 코드 이동
